requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==4.1.3
    # via
    #   -r requirements.in
    #   django-filter
    #   django-stubs-ext
    #   django-taggit
    #   djangorestframework
    #   drf-yasg
django-filter==22.1
    # via -r requirements.in
django-stubs-ext==0.7.0
    # via -r requirements.in
django-taggit==3.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg[validation]==1.21.4
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonschema==4.17.0
    # via swagger-spec-validator
markdown==3.4.1
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
packaging==21.3
    # via drf-yasg
pillow==9.3.0
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pyrsistent==0.19.2
    # via jsonschema
pytz==2022.6
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via swagger-spec-validator
requests==2.28.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
sqlparse==0.4.3
    # via django
swagger-spec-validator==3.0.3
    # via drf-yasg
typing-extensions==4.4.0
    # via
    #   django-stubs-ext
    #   swagger-spec-validator
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via requests
